Address comment and rename
