+++ consensus/polybft/ibft_multicast_noop.go
@@ -4,5 +4,5 @@ package polybft
 
 import ibftProto "github.com/0xPolygon/go-ibft/messages/proto"
 
-func (p *Polybft) customMulticastHandler(msg *ibftProto.Message) {
+func (p *Polybft) ibftMsgMulticast(msg *ibftProto.Message) {
 }
